Update to new Reloaded GitHub action

+++ p4gpc.tinyadditions/Publish.ps1
@@ -46,6 +46,10 @@
 
     If this is true, you should set UseGitHubDelta, UseGameBananaDelta, UseNuGetDelta or equivalent to true.
 
+.PARAMETER MetadataFileName
+    Default: Sewer56.Update.ReleaseMetadata.json
+    Name of the release metadata file used to download the delta package.
+
 .PARAMETER UseGitHubDelta
     Default: $False
     If true, sources the last version of the package to publish from GitHub.
@@ -70,6 +74,11 @@
     [Use if UseGitHubDelta is true]
     Allows you to specify a Wildcard pattern (e.g. *Update.zip) for the file to be downloaded.
     This is a fallback used in cases no Release Metadata file can be found.
+
+.PARAMETER GitHubInheritVersionFromTag
+    [Use if UseGitHubDelta is true]
+    Uses version determined from release tag (in GitHub Releases) as opposed to the 
+    Release Metadata file in latest release.
 
 .PARAMETER GameBananaItemId
     [Use if UseGameBananaDelta is true]
@@ -111,6 +120,17 @@
 
     Publishes a package that can be uploaded to GameBanana.
 
+.PARAMETER Build
+    Default: $True
+
+    Whether the project should be built.
+    Setting this to false lets you use the publish part of the script standalone in a non .NET environment.
+	
+.PARAMETER RemoveExe
+    Default: $True
+
+    Removes executables from build output. Useful when performing R2R Optimisation.
+
 .EXAMPLE
   .\Publish.ps1 -ProjectPath "Reloaded.Hooks.ReloadedII/Reloaded.Hooks.ReloadedII.csproj" -PackageName "Reloaded.Hooks.ReloadedII" -PublishOutputDir "Publish/ToUpload"
 
@@ -126,27 +146,31 @@ param (
     $IsPrerelease=$False, 
     $MakeDelta=$False, 
     $ChangelogPath="",
+    $Build=$True,
     $BuildR2R=$False,
-    
+    $RemoveExe = $True,
+	
     ## => User Config <= ## 
-    $ProjectPath = "p4gpc.tinyadditions.csproj",
+    $ProjectPath = "p4gpc.tinyadditions/p4gpc.tinyadditions.csproj",
     $PackageName = "p4gpc.tinyadditions",
     $PublishOutputDir = "Publish/ToUpload",
 
     ## => User: Delta Config
     # Pick one and configure settings below.
-    $UseGitHubDelta = $False,
+    $MetadataFileName = "Sewer56.Update.ReleaseMetadata.json",
+    $UseGitHubDelta = $False, # GitHub Releases
     $UseGameBananaDelta = $False,
     $UseNuGetDelta = $False,
 
-    $GitHubUserName = "AnimatedSwine37",
-    $GitHubRepoName = "p4gpc.tinyadditions",
-    $GitHubFallbackPattern = "", # For migrating from legacy.
+    $GitHubUserName = "AnimatedSwine37", # Name of the GitHub user where the mod is contained
+    $GitHubRepoName = "p4gpc.tinyadditions", # Name of the GitHub repo where the mod is contained
+    $GitHubFallbackPattern = "", # For migrating from legacy build script.
+    $GitHubInheritVersionFromTag = $True, # Uses version determined from release tag as opposed to metadata file in latest release.
 
-    $GameBananaItemId = 0, # From mod page URL.
+    $GameBananaItemId = 345639, # From mod page URL.
 
     $NuGetPackageId = "p4gpc.tinyadditions",
-    $NuGetFeedUrl = "",
+    $NuGetFeedUrl = "http://packages.sewer56.moe:5000/v3/index.json",
     $NuGetAllowUnlisted = $False,
 
     ## => User: Publish Config
@@ -156,8 +180,8 @@ param (
 )
 
 ## => User: Publish Output
-$publishBuildDirectory = "Publish/Builds/CurrentVersion"      # Build directory for current version of the mod.
-$deltaDirectory = "Publish/Builds/LastVersion"                # Path to last version of the mod.
+$publishBuildDirectory = "p4gpc.tinyadditions/Publish/Builds/CurrentVersion"      # Build directory for current version of the mod.
+$deltaDirectory = "p4gpc.tinyadditions/Publish/Builds/LastVersion"                # Path to last version of the mod.
 
 $PublishGenericDirectory = "$PublishOutputDir/Generic"        # Publish files for any target not listed below.
 $PublishNuGetDirectory   = "$PublishOutputDir/NuGet"          # Publish files for NuGet
@@ -165,10 +189,11 @@ $PublishGameBananaDirectory = "$PublishOutputDir/GameBanana"  # Publish files fo
 
 ## => User Config <= ## 
 # Tools
-$reloadedToolsPath = "./Publish/Tools/Reloaded-Tools"    # Used to check if tools are installed.
-$updateToolsPath   = "./Publish/Tools/Update-Tools"      # Used to check if update tools are installed.
+$reloadedToolsPath = "./p4gpc.tinyadditions/Publish/Tools/Reloaded-Tools"    # Used to check if tools are installed.
+$updateToolsPath   = "./p4gpc.tinyadditions/Publish/Tools/Update-Tools"      # Used to check if update tools are installed.
 $reloadedToolPath = "$reloadedToolsPath/Reloaded.Publisher.exe"  # Path to Reloaded publishing tool.
 $updateToolPath   = "$updateToolsPath/Sewer56.Update.Tool.dll" # Path to Update tool.
+$changelogFullPath = [System.IO.Path]::GetFullPath($ChangelogPath)
 
 ## => Script <= ##
 # Set Working Directory
@@ -178,7 +203,9 @@ Split-Path $MyInvocation.MyCommand.Path | Push-Location
 # Convert Booleans
 $IsPrerelease = [bool]::Parse($IsPrerelease)
 $MakeDelta = [bool]::Parse($MakeDelta)
+$Build = [bool]::Parse($Build)
 $BuildR2R = [bool]::Parse($BuildR2R)
+$RemoveExe = [bool]::Parse($RemoveExe)
 $UseGitHubDelta = [bool]::Parse($UseGitHubDelta)
 $UseGameBananaDelta = [bool]::Parse($UseGameBananaDelta)
 $UseNuGetDelta = [bool]::Parse($UseNuGetDelta)
@@ -186,28 +213,29 @@ $NuGetAllowUnlisted = [bool]::Parse($NuGetAllowUnlisted)
 $PublishGeneric = [bool]::Parse($PublishGeneric)
 $PublishNuGet = [bool]::Parse($PublishNuGet)
 $PublishGameBanana = [bool]::Parse($PublishGameBanana)
-
-Write-Host "IsPrerelease $IsPrerelease, MakeDelta: $MakeDelta, Changelog: $ChangelogPath"
+$GitHubInheritVersionFromTag = [bool]::Parse($GitHubInheritVersionFromTag)
+$TempDirectory = [System.IO.Path]::GetTempPath() + [System.IO.Path]::GetRandomFileName()
+$TempDirectoryBuild = "$TempDirectory/build"
 
 function Get-Tools {
     # Download Tools (if needed)
     $ProgressPreference = 'SilentlyContinue'
     if (-not(Test-Path -Path $reloadedToolsPath -PathType Any)) {
         Write-Host "Downloading Reloaded Tools"
-        Invoke-WebRequest -Uri "https://github.com/Reloaded-Project/Reloaded-II/releases/latest/download/Tools.zip" -OutFile "$env:TEMP/Tools.zip"
-        Expand-Archive -LiteralPath "$env:TEMP/Tools.zip" -DestinationPath $reloadedToolsPath
+        Invoke-WebRequest -Uri "https://github.com/Reloaded-Project/Reloaded-II/releases/latest/download/Tools.zip" -OutFile "$TempDirectory/Tools.zip"
+        Expand-Archive -LiteralPath "$TempDirectory/Tools.zip" -DestinationPath $reloadedToolsPath
 
         # Remove Items
-        Remove-Item "$env:TEMP/Tools.zip" -ErrorAction SilentlyContinue
+        Remove-Item "$TempDirectory/Tools.zip" -ErrorAction SilentlyContinue
     }
 
     if ($MakeDelta -and -not(Test-Path -Path $updateToolsPath -PathType Any)) {
         Write-Host "Downloading Update Library Tools"
-        Invoke-WebRequest -Uri "https://github.com/Sewer56/Update/releases/latest/download/Sewer56.Update.Tool.zip" -OutFile "$env:TEMP/Sewer56.Update.Tool.zip"
-        Expand-Archive -LiteralPath "$env:TEMP/Sewer56.Update.Tool.zip" -DestinationPath $updateToolsPath
+        Invoke-WebRequest -Uri "https://github.com/Sewer56/Update/releases/latest/download/Sewer56.Update.Tool.zip" -OutFile "$TempDirectory/Sewer56.Update.Tool.zip"
+        Expand-Archive -LiteralPath "$TempDirectory/Sewer56.Update.Tool.zip" -DestinationPath $updateToolsPath
 
         # Remove Items
-        Remove-Item "$env:TEMP/Sewer56.Update.Tool.zip" -ErrorAction SilentlyContinue    
+        Remove-Item "$TempDirectory/Sewer56.Update.Tool.zip" -ErrorAction SilentlyContinue    
     }
 }
 
@@ -222,8 +250,8 @@ function Build {
     dotnet clean $ProjectPath
 
     if ($BuildR2R) {
-        dotnet publish $ProjectPath -c Release -r win-x86 --self-contained false -o "$publishBuildDirectory/x86" /p:PublishReadyToRun=true
-        dotnet publish $ProjectPath -c Release -r win-x64 --self-contained false -o "$publishBuildDirectory/x64" /p:PublishReadyToRun=true
+        dotnet publish $ProjectPath -c Release -r win-x86 --self-contained false -o "$publishBuildDirectory/x86" /p:PublishReadyToRun=true /p:OutputPath="$TempDirectoryBuild/x86"
+        dotnet publish $ProjectPath -c Release -r win-x64 --self-contained false -o "$publishBuildDirectory/x64" /p:PublishReadyToRun=true /p:OutputPath="$TempDirectoryBuild/x64"
 
         # Remove Redundant Files
         Move-Item -Path "$publishBuildDirectory/x86/ModConfig.json" -Destination "$publishBuildDirectory/ModConfig.json" -ErrorAction SilentlyContinue
@@ -232,11 +260,15 @@ function Build {
         Remove-Item "$publishBuildDirectory/x64/ModConfig.json" -ErrorAction SilentlyContinue
     }
     else {
-        dotnet publish $ProjectPath -c Release --self-contained false -o "$publishBuildDirectory"
+        dotnet publish $ProjectPath -c Release --self-contained false -o "$publishBuildDirectory" /p:OutputPath="$TempDirectoryBuild"
     }
 
     # Cleanup Unnecessary Files
-    Get-ChildItem $publishBuildDirectory -Include *.exe -Recurse | Remove-Item -Force -Recurse
+    Remove-Item $TempDirectoryBuild -Recurse -ErrorAction SilentlyContinue
+	if ($RemoveExe) {
+        Get-ChildItem $publishBuildDirectory -Include *.exe -Recurse | Remove-Item -Force -Recurse
+	}
+	
     Get-ChildItem $publishBuildDirectory -Include *.pdb -Recurse | Remove-Item -Force -Recurse
     Get-ChildItem $publishBuildDirectory -Include *.xml -Recurse | Remove-Item -Force -Recurse
 }
@@ -245,10 +277,10 @@ function Get-Last-Version {
     
     Remove-Item $deltaDirectory -Recurse -ErrorAction SilentlyContinue
     New-Item $deltaDirectory -ItemType Directory -ErrorAction SilentlyContinue
-    $arguments = "DownloadPackage --extract --outputpath `"$deltaDirectory`" --allowprereleases `"$IsPrerelease`""
+    $arguments = "DownloadPackage --extract --outputpath `"$deltaDirectory`" --allowprereleases `"$IsPrerelease`" --metadatafilename `"$MetadataFileName`""
 	
     if ($UseGitHubDelta) {
-        $arguments += " --source GitHub --githubusername `"$GitHubUserName`" --githubrepositoryname `"$GitHubRepoName`" --githublegacyfallbackpattern `"$GitHubFallbackPattern`""
+        $arguments += " --source GitHub --githubusername `"$GitHubUserName`" --githubrepositoryname `"$GitHubRepoName`" --githublegacyfallbackpattern `"$GitHubFallbackPattern`" --githubinheritversionfromtag `"$GitHubInheritVersionFromTag`""
     }
     elseif ($UseNuGetDelta) {
         $arguments += " --source NuGet --nugetpackageid `"$NuGetPackageId`" --nugetfeedurl `"$NuGetFeedUrl`" --nugetallowunlisted `"$NuGetAllowUnlisted`""
@@ -268,7 +300,7 @@ function Get-Common-Publish-Args {
 	
 	$arguments = "--modfolder `"$publishBuildDirectory`" --packagename `"$PackageName`""
 	if ($ChangelogPath) {
-        $arguments += " --changelogpath `"$ChangelogPath`""
+        $arguments += " --changelogpath `"$changelogFullPath`""
 	}
 	
 	if ($AllowDeltas -and $MakeDelta) {
@@ -315,6 +347,7 @@ function Cleanup {
 }
 
 # Build & Publish
+New-Item $TempDirectory -ItemType Directory -ErrorAction SilentlyContinue
 Cleanup
 Get-Tools
 
@@ -323,8 +356,10 @@ if ($MakeDelta) {
     Get-Last-Version
 }
 
-Write-Host "Building Mod"
-Build
+if ($Build) {
+    Write-Host "Building Mod"
+    Build    
+}
 
 if ($PublishGeneric) {
     Write-Host "Publishing Mod for Default Target"
@@ -340,6 +375,9 @@ if ($PublishGameBanana) {
     Write-Host "Publishing Mod for GameBanana Target"
     Publish-GameBanana
 }
+
+# Remove Temp Folder
+Remove-Item $TempDirectory -Recurse -ErrorAction SilentlyContinue
 
 # Restore Working Directory
 Write-Host "Done."
